publish-tdl.yml: Move to centos:7 devtoolset-9
Move from prebuild/centos7-devtoolset7 to centos:7 with devtoolset-9
(the same version Holy Build Box uses, which is used in prebuilt-tdlib).
This image has a newer gcc version, hopefully fixing a compilation error
that was present with gcc 7.

Note: We don't link libstdc++ statically to reduce the binary size
(about 1.1M with statically linked libstdc++) and consequently the npm
package size. This is the same as before this commit, since
prebuildify-cross links libstdc++ dynamically as well.