addon: Use a dedicated thread per client
Removes the reuse of the uv threadpool in the old tdjson interface. This
also lifts the limitation of max UV_THREADPOOL_SIZE clients.

The timeout argument is moved from the receive binding to createClient.

bun seems to work after the removal of Napi::AsyncWorker, though the
stability might not be top notch. "Other JavaScript runtimes" section is
added to the README.