tdl-install-types: Update the generator
Added $FunctionResultByName, $FunctionInputByName special object types
that are now used in Invoke and Execute instead of chains of extends +
conditional types (should be faster to typecheck, also less repetitive,
and can be used in flow). Also added $FunctionName and $SyncFunctionName
that are used in "extends" in the type parameters of Invoke and Execute
respectively.