prebuilt-tdlib update
The publishing process of prebuilt-tdlib is heavily changed, everything
is now in the publish.js script (check-prebuilds.js and
write-tdlib-json.js are subsumed), which prepares the prebuilt-tdlib
package (updates its name and version, removes private) and publishes
it.

It is no longer necessary to specify the full padded npm version to run
the prebuilt-tdlib workflow. The input parameter is changed from
npm-version to npm-patch (accepts a single number).

The name of prebuilt-tdlib/package.json in this repository is now
prebuilt-tdlib-dev, which can allow to install prebuilt-tdlib as a dev
dependency here in the future.

The prebuilt-tdlib package is now published with the tdlib field in
package.json (contains "commit" and "version" properties), which should
enable to run convenient queries like `npm info prebuilt-tdlib
tdlib.commit`.